pyproject.toml
[tool.poetry]
name = "music_flash_cards"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.scripts]
music_flash_cards = "music_flash_cards.music_note_flash_cards:main"

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.0.1"
PyAudio = "^0.2.11"
typer = "^0.3.2"
librosa = "^0.8.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
